feat: split ViewCounter into Suspense boundary + client component, remove blocking getViews
